fix: force close on lifecycle (ketika spam bolak balik page 😃😩)

+++ app/src/main/java/com/entsh104/highking/ui/cust/profile/ProfileFragment.kt
@@ -14,7 +14,9 @@ import android.widget.Toast
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
 import com.entsh104.highking.R
@@ -22,6 +24,7 @@ import com.entsh104.highking.data.source.local.SharedPreferencesManager
 import com.entsh104.highking.data.source.remote.RetrofitClient
 import com.entsh104.highking.databinding.FragmentCustProfileBinding
 import com.entsh104.highking.ui.auth.AuthActivity
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 class ProfileFragment : Fragment() {
@@ -52,17 +55,23 @@ class ProfileFragment : Fragment() {
         binding.btnLogout.setOnClickListener {
             val token = prefs.getToken().toString()
             viewLifecycleOwner.lifecycleScope.launch {
-                val responseLogout = userRepository.logoutUser(token)
-                if (responseLogout.isSuccess) {
-                    Toast.makeText(requireContext(), "Logout Berhasil", Toast.LENGTH_SHORT).show()
-                    prefs.clear()
-                    val intent = Intent(activity, AuthActivity::class.java)
-                    startActivity(intent)
-                    activity?.finish()
-                } else {
-                    Toast.makeText(requireContext(), "Logout failed", Toast.LENGTH_SHORT).show()
+                lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                    delay(500)
+                    if (viewLifecycleOwner.lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)) {
+                        val responseLogout = userRepository.logoutUser(token)
+                        if (responseLogout.isSuccess) {
+                            Toast.makeText(requireContext(), "Logout Berhasil", Toast.LENGTH_SHORT)
+                                .show()
+                            prefs.clear()
+                            val intent = Intent(activity, AuthActivity::class.java)
+                            startActivity(intent)
+                            activity?.finish()
+                        } else {
+                            Toast.makeText(requireContext(), "Logout failed", Toast.LENGTH_SHORT)
+                                .show()
+                        }
+                    }
                 }
-
             }
         }
 
@@ -97,20 +106,29 @@ class ProfileFragment : Fragment() {
         builder.setMessage("Are you sure you want to upload this photo?")
         builder.setPositiveButton("Yes") { dialog, which ->
             viewLifecycleOwner.lifecycleScope.launch {
-                val response = userRepository.uploadPhoto(
-                    requireContext().contentResolver,
-                    requireContext().cacheDir,
-                    imageUri
-                )
-                if (response.isSuccess) {
-                    Toast.makeText(
-                        requireContext(),
-                        "Photo uploaded successfully",
-                        Toast.LENGTH_SHORT
-                    ).show()
-                    fetchUserProfile()
-                } else {
-                    Toast.makeText(requireContext(), "Failed to upload photo", Toast.LENGTH_SHORT).show()
+                lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                    delay(500)
+                    if (viewLifecycleOwner.lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)) {
+                        val response = userRepository.uploadPhoto(
+                            requireContext().contentResolver,
+                            requireContext().cacheDir,
+                            imageUri
+                        )
+                        if (response.isSuccess) {
+                            Toast.makeText(
+                                requireContext(),
+                                "Photo uploaded successfully",
+                                Toast.LENGTH_SHORT
+                            ).show()
+                            fetchUserProfile()
+                        } else {
+                            Toast.makeText(
+                                requireContext(),
+                                "Failed to upload photo",
+                                Toast.LENGTH_SHORT
+                            ).show()
+                        }
+                    }
                 }
             }
         }
@@ -138,27 +156,36 @@ class ProfileFragment : Fragment() {
 
     private fun fetchUserProfile() {
         viewLifecycleOwner.lifecycleScope.launch {
-            binding.progressBar.visibility = View.VISIBLE
-            binding.scrollViewContent.visibility = View.GONE
+            lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                delay(500)
+                if (viewLifecycleOwner.lifecycle.currentState.isAtLeast(Lifecycle.State.STARTED)) {
+                    binding.progressBar.visibility = View.VISIBLE
+                    binding.scrollViewContent.visibility = View.GONE
 
-            val result = userRepository.getCurrentUser()
-            if (result.isSuccess) {
-                val user = result.getOrNull()
-                user?.let {
-                    Log.d("ProfileFragment", "User: $it")
-                    binding.textViewUsername.text = "@" + it.username
-                    binding.textViewName.text = it.username
-                    binding.textViewEmail.text = it.email
-                    binding.textViewPhone.text = it.phone_number
-                    Glide.with(this@ProfileFragment).load("${it.image_url}")
-                        .into(binding.imageViewProfile)
-                    binding.progressBar.visibility = View.GONE
-                    binding.scrollViewContent.visibility = View.VISIBLE
+                    val result = userRepository.getCurrentUser()
+                    if (result.isSuccess) {
+                        val user = result.getOrNull()
+                        user?.let {
+                            Log.d("ProfileFragment", "User: $it")
+                            binding.textViewUsername.text = "@" + it.username
+                            binding.textViewName.text = it.username
+                            binding.textViewEmail.text = it.email
+                            binding.textViewPhone.text = it.phone_number
+                            Glide.with(this@ProfileFragment).load("${it.image_url}")
+                                .into(binding.imageViewProfile)
+                            binding.progressBar.visibility = View.GONE
+                            binding.scrollViewContent.visibility = View.VISIBLE
+                        }
+                    } else {
+                        Toast.makeText(
+                            requireContext(),
+                            "Failed to load profile",
+                            Toast.LENGTH_SHORT
+                        )
+                            .show()
+                        binding.progressBar.visibility = View.GONE
+                    }
                 }
-            } else {
-                Toast.makeText(requireContext(), "Failed to load profile", Toast.LENGTH_SHORT)
-                    .show()
-                binding.progressBar.visibility = View.GONE
             }
         }
     }
